% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-metrics.R
\name{generate_confusion_matrix}
\alias{generate_confusion_matrix}
\title{Generate a confusion matrix from a random forest as a tibble}
\usage{
generate_confusion_matrix(rf)
}
\arguments{
\item{rf}{A randomForest object.}
}
\value{
Tibble.
}
\description{
Generate a confusion matrix from a random forest as a tibble
}
