% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-metrics.R
\name{generate_metrics}
\alias{generate_metrics}
\title{Generate a tibble of metrics from a random forest}
\usage{
generate_metrics(rf)
}
\arguments{
\item{rf}{A randomForest object.}
}
\value{
Tibble.
}
\description{
Generate a tibble of metrics from a random forest
}
