% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-roc.R
\name{generate_roc}
\alias{generate_roc}
\title{Generate a ROC curve from the training data of a randomForest model}
\usage{
generate_roc(rf)
}
\arguments{
\item{rf}{A model created with the \verb{randomForest} package.}
}
\value{
A \verb{ggplot2} plot.
}
\description{
In a real-life application, we wouldn't be using training data to plot our
ROC curve, as random forests greatly overfit on their training data. Since
this isn't a serious exercise in machine learning, we can get away with it.
}
