% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-metrics.R
\name{generate_metrics}
\alias{generate_metrics}
\title{Generate a tibble of metrics from a random forest}
\usage{
generate_metrics(rf, positive)
}
\arguments{
\item{rf}{A model created with the \verb{randomForest} package.}

\item{positive}{Character. The value of the positive case in the binomial
classification.}
}
\value{
Tibble.
}
\description{
Generate a tibble of metrics from a random forest
}
